src/requirements.txt
wikipedia==1.4.0
nltk==3.5
requests==2.24.0
pyjokes==0.6.0
pyttsx3==2.90
SpeechRecognition==3.8.1
PyAutoGUI==0.9.52
Pillow==8.3.0
